+++ rust/stackablectl/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add new argument `--release` that allows installing a specific version of a demo or stack ([#340]).
+
+### Removed
+
+- Remove argument `--offline` that was not implemented yet ([#340]).
+
+[#340]: https://github.com/stackabletech/stackable-cockpit/pull/340
+
 ## [24.11.1] - 2024-11-20
 
 ### Added